move Transcoder to Transcoder.Server
